Add argparser

+++ include/lz77_compression.h
@@ -37,7 +37,7 @@ namespace sfic{
         void shift(RawData const& data, unsigned int quantity);
         // given an index in the search buffer , it returns the length
         // the longest match from that position
-        unsigned int searchFromIndex(RawData const& data, std::deque<int>::const_iterator i) const;
+        unsigned int searchFromIndex(RawData const& data, std::deque<ByteType>::const_iterator i) const;
         // converts an integer to a string efficiently
         static std::string toString(unsigned int i);
         // the buffer size fixes the physical size of the search buffer
@@ -62,7 +62,7 @@ namespace sfic{
         // where characters that have already been seen
         // are saved and discarted on buffer fill
         // implemented as a deque for efficiency since
-        // we can use it as a circular buffer 
+        // we can use it as a circular buffer
         std::deque<ByteType> searchBuffer;
